+++ Example/ChildVC/FirstViewController.m
@@ -50,7 +50,7 @@
 
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath {
     if (isDebug) {
-        NSLog(@"点击了控制器1单元格的%li",indexPath.row);
+        NSLog(@"点击了控制器1单元格的%li",(long)indexPath.row);
     }    
     TestViewController *testVC = [TestViewController new];
     [self.navigationController pushViewController:testVC animated:YES];
